requirements.txt
asgiref==3.7.2
Django==4.2.6
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Markdown==3.5
PyJWT==1.7.1
PyMySQL==1.1.0
pytz==2023.3.post1
sqlparse==0.4.4
